Optimize scheduler for IO heavy tasks
The simple round robin scheduler was changed to de-prioritize a task
if a thread is waiting on a lock. I chose a value of 8, which means
that 8 cycles of a task will be skipped if it is running lock. This
number might not be optimal for all tasks.

Basic Round Robin: *4:47*
Optimized Round Robin: *0:44*

This is a pretty clear optimization. Also, fun fact: Both versions of
these are slower with `O2` and `-accel kvm` than without. This isn't
intuitive, but makes sense. Since the code to output text will run
faster, a thread will complete its work faster. That also means that
the queue will fill up faster. Ultimately, threads will spend more
time of the program waiting on a lock than they usually will.