Add some basic file system support.
I gave up on trying to fix that memory leak fully because of a weird
bug. Now there is a lisp function to read the text from a file:

(read "initrd/hello.txt")

By the way, I also added support for exec_node with a string, which
was really easy, and I don't know why I didn't do that already.