Fix static pointers in ELF loader.
Before, when I tried to store the ascii art rubber duck in a static
variable, the wrong pointer would be passed to the system call, and it
wouldn't be able to be printed. This is because the pointer is given
relative to the start of the section of the ELF file that is being
executed, rather than its actual memory location. This can be
sidestepped by adding a function to translate out-of-bounds memory
addresses to be relative to the file boundary.

Landmines (Potential future bugs):
- If I add a malloc system call, the heap-allocated memory will be out
of bounds of the ELF, leading to it being erroneously translated, and
causing a bug. There are a couple possible fixes:
  - After setting up paging, we switch the malloc implementation to
  work in a different way. This switch happens after we copy over the
  file system. Thus, we can keep track of whether a section of memory
  is heap allocated, and not transfer it over. This will go wrong if
  we ever have a file system that isn't a ram disk.
  - Make a table of allocations that are provided to the operating
  system, and check the table to see if an allocation is from
  malloc. This would work, but would add considerable overhead to
  every single memory allocation on the system.
  - Actually fix the issue. There's probably some GCC flag that I
  don't know about that fixes this, but I don't know of one.
  - Only transform memory addresses that are below a certain (very
  small) memory address. The problem with this is that it assumes that
  all of our binaries will be around 5kb, and all incorrect memory
  addresses will be within this range. If we ever make a bigger
  binary, then we're in trouble.

I'm going to look into option 2.