Handle the global offset table properly.
After checking the relevant Wikipedia article, I realized that the
correct solution to the problem in the last commit was to offset the
data in the Global Offset Table when the program is being loaded. We
also use one of the padding bytes in the ELF file header to represent
whether we have already offset the table or not. We need to do this
because all of the ELF files are stored in read/write memory, and we
don't have multiple copies. If we didn't store this data, we would
offset the pointers multiple times, and it would cause bugs.