Implement and use the return value of main
Usually, the main function in C returns an integer value representing
whether the intended operation succeeded (0) or failed (any other
number). This commit implements this and returns it when you use the
lisp exec method. This is unimportant and might end up being removed,
but it exists. Also, the printf_test.c was modified to return 123, so
you can test this with (exec "initrd/test.elf")