Some codespaces changes
I wanted to work on this in a github codespace. I had to fix some things to get it to boot, presumably because codespaces runs on Ubuntu, which has older compiler versions. In doing this, a bug that I've been stuck on for a while has been fixed. I don't know if this will transfer over, but hopefully it will :)